*** empty log message ***

+++ magick/zoom.c
@@ -1081,7 +1081,8 @@ MagickExport Image *ResizeImage(Image *image,const unsigned int columns,
     Resize image.
   */
   quantum=0;
-  if ((columns*(image->rows+rows)) < (rows*(image->columns+columns)))
+  if ((size_t) (columns*(image->rows+rows)) <
+      (size_t) (rows*(image->columns+columns)))
     {
       source_image=CloneImage(image,columns,image->rows,True,exception);
       if (source_image == (Image *) NULL)
